requirements.txt
Flask-RESTful
Flask
foursquare
pycountry
python-dotenv